+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Features:
 
   - Allow primary button classes to be overridden (#183, @tanimichi)
   - Allow to pass a block to select helper in Rails >= 4.1.0 (#229, @doits)
+  - Allow inline form groups, to group several elements in one line
 
 ## 2.3.0 (2015-02-17)